Refactor: fix args
# 65
