+++ config/BERT4Rec_config.yaml
@@ -1,0 +1,22 @@
+seed : 0
+device: cpu
+model: BERT4Rec
+
+model_args:
+  BERT4Rec:
+    num_epochs: 3
+    batch_size: 128
+    lr: 1e-3
+    mask_prob: 0.15
+    max_len: 50
+    hidden_units: 50
+    num_heads: 1
+    num_layers: 2
+    dropout_rate: 0.5
+    num_workers: 1
+
+
+dataset :
+  data_path : train/
+  output_path : saved/output
+  preprocessing_path : saved/preprocessed/
