--- src/BERT4Rec/BERT4Rec.yaml
@@ -1,75 +1,0 @@
-V0:
-  config_ver: 0
-  model: BERT4Rec
-  data_path: /data/ephemeral/home/data/train/
-  model_save_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  submit_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  seed: 42
-  parameters:
-    num_epochs: 3
-    batch_size: 128
-    lr: 1e-3
-    mask_prob: 0.15
-    max_len: 50
-    hidden_units: 50
-    num_heads: 1
-    num_layers: 2
-    dropout_rate: 0.5
-    num_workers: 1
-
-V1:
-  config_ver: 1
-  model: BERT4Rec
-  data_path: /data/ephemeral/home/data/train/
-  model_save_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  submit_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  seed: 42
-  parameters:
-    num_epochs: 10
-    batch_size: 128
-    lr: 1e-3
-    mask_prob: 0.15
-    max_len: 50
-    hidden_units: 50
-    num_heads: 1
-    num_layers: 2
-    dropout_rate: 0.5
-    num_workers: 1
-
-V2:
-  config_ver: 2
-  model: BERT4Rec
-  data_path: /data/ephemeral/home/data/train/
-  model_save_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  submit_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  seed: 42
-  parameters:
-    num_epochs: 5
-    batch_size: 128
-    lr: 1e-3
-    mask_prob: 0.15
-    max_len: 50
-    hidden_units: 30
-    num_heads: 1
-    num_layers: 2
-    dropout_rate: 0.5
-    num_workers: 1
-
-V3:
-  config_ver: 3
-  model: BERT4Rec
-  data_path: /data/ephemeral/home/data/train/
-  model_save_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  submit_path: /data/ephemeral/home/kje/BERT4Rec/src/
-  seed: 42
-  parameters:
-    num_epochs: 10
-    batch_size: 1024
-    lr: 1e-5
-    mask_prob: 0.15
-    max_len: 50
-    hidden_units: 50
-    num_heads: 1
-    num_layers: 2
-    dropout_rate: 0.5
-    num_workers: 1
